requirements.txt
django
langchain
langchain-openai
python-dotenv
tiktoken